Fix gpg invocation when non-root
The export command here writing the key into /etc/apt/trusted.gpg.d
didn't work because the `gpg --export` was under sudo, but the shell
redirection `> /etc/....` wasn't, so if non-root you'd get a permission
denied error.  (And if you *are* root, the `sudo` is pointless).

But there's really no reason to run *any* of the gpg commands as root,
so this fixes it all to:

1. Remove the `sudo`s from gpg invocations.
2. Change the shell redirection to a `| sudo tee ...` so that only the
   output, but not the command, necessarily runs as root.